requirements.txt
flask==2.1.2
pycryptodomex==3.14.1
gevent==21.12.0
requests==2.27.1
sqlalchemy==1.4.36
psycopg2-binary==2.9.3
flask-sqlalchemy==2.5.1
flask_restful==0.3.9
python-dotenv==0.20.0
flake8==4.0.1
pydantic==1.9.1
email-validator==1.2.1
setuptools==63.2.0
flasgger==0.9.5
